Completing the Menu task
