+++ src/app/app-routing.module.ts
@@ -1,21 +1,22 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
+import { MenuComponent } from './components/menu/menu.component';
+// import { CustomersComponent } from './components/customers/customers.component';
 import { RegisterComponent } from './register/register.component';
 import { LoginComponent } from './login/login.component';
 import { HomeComponent } from './home/home.component';
 import { ProfileComponent } from './profile/profile.component';
 import { UserComponent } from './costumer/user/user.component';
-
 import { AdminComponent } from './costumer/admin/admin.component';
 
 const routes: Routes = [
-  { path: 'home', component: HomeComponent },
   { path: 'login', component: LoginComponent },
   { path: 'register', component: RegisterComponent },
   { path: 'profile', component: ProfileComponent },
+  // {path:"customers", component: CustomersComponent},
+  {path:"menu", component:MenuComponent},
   { path: 'user', component: UserComponent },
-
   { path: 'admin', component: AdminComponent },
   { path: '', redirectTo: 'home', pathMatch: 'full' }
 ];
