+++ src/app/app.component.html
@@ -1,41 +1,2 @@
-<div id="app">
-  <nav class="navbar" style="background-color: #e3f2fd;">
-    <a href="#" class="navbar-brand">Delivery food</a>
-    <ul class="navbar mr-auto" routerLinkActive="active">
-      <li class="nav">
-        <a href="/home" class="nav-link" routerLink="home">Home </a>
-      </li>
-      <li class="nav" *ngIf="showAdminBox">
-        <a href="/admin" class="nav-link" routerLink="admin">Admin</a>
-      </li>
-
-      <li class="nav">
-        <a href="/user" class="nav-link" *ngIf="isLoggedIn" routerLink="user">User</a>
-      </li>
-    </ul>
-
-    <ul class="navbar ml-auto" *ngIf="!isLoggedIn">
-      <li class="nav">
-        <a href="/register" class="nav-link" routerLink="register">Sign Up</a>
-      </li>
-      <li class="nav">
-        <a href="/login" class="nav-link" routerLink="login">Login</a>
-      </li>
-    </ul>
-
-    <ul class="navbar ml-auto" *ngIf="isLoggedIn">
-      <li class="nav">
-        <a href="/profile" class="nav-link" routerLink="profile">{{ firstName }}</a>
-        <a href="/profile" class="nav-link" routerLink="profile">{{ lastName }}</a>
-      </li>
-
-      <li class="nav">
-        <a href class="nav-link" (click)="logout()">LogOut</a>
-      </li>
-    </ul>
-  </nav>
-
-  <div class="container">
-    <router-outlet></router-outlet>
-  </div>
-</div>
+    <app-home></app-home>
+    <router-outlet></router-outlet>
